generic/260: don't fail for certain fstrim ops on btrfs
We have always failed generic/260, because it tests to see if the file
system will reject a trim range that is above the reported fs size.

However for btrfs we will happily remap logical byte offsets within the
file system, so you can end up with bye offsets past the end of the
reported end of the file system.  Thus we do not fail these weird
ranges.  We also don't have the concept of allocation groups, so the
other test that tries to catch overflow doesn't apply to us either.

Fix this by simply using an offset that will fail (once a related
kernel path is applied) for btrfs.  This will allow us to test the
different overflow cases that do apply to btrfs, and not muddy up
test results by giving us a false negative for the cases that do not
apply to btrfs.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>
Acked-by: David Sterba <dsterba@suse.cz>
Signed-off-by: Eryu Guan <guaneryu@gmail.com>