generic/274: require no compress
We hit spurious errors with generic/274 with compression on because we
attempt to fill up the disk with small writes, and these writes end up
taking up metadata space instead of data space.  Thus when we go to
write into the preallocated area we get an ENOSPC, but from the metadata
side and not the data side.  Simply skip this test if we have
compression enabled.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>
Reviewed-by: Eryu Guan <guaneryu@gmail.com>
Signed-off-by: Eryu Guan <guaneryu@gmail.com>