btrfs/011: continue the test if we failed to cancel the replace
Btrfs/011 test case has exposed quite some bugs in the past, but it also
has some tendency to cause false alert, as its way testing replace
cancel require the replace to be still running when we send the cancel
request.

But on a lot of cases, the replace can finish way faster than the wait
time, and cause false alert.

Commit fa85aa64 ("btrfs/011: Fill the fs to ensure we have enough data for
dev-replace") tries to address the problem by filling the fs, but there
is still no guarantee.

Although there is still some discussion on how to properly solve the
problem, there is one thing sure that we should continue the test
instead of abort, if the replace cancel failed.

A quick abort caused by finished replace will leave other profiles
untested, reducing the coverage.

This patch will still mark the test failed for a finished replace, but
at least ensure we have run the test for all the profiles.

Signed-off-by: Qu Wenruo <wqu@suse.com>
Reviewed-by: Eryu Guan <guaneryu@gmail.com>
Signed-off-by: Eryu Guan <guaneryu@gmail.com>